🧹 refactor: changed response channels into bounded channels
- response channel sends happen in the async tasks
  so no real reasons they should be unbounded
- added err handling back to the connection
