+++ src/tui/builder.rs
@@ -1,15 +1,5 @@
 use super::*;
 
-#[derive(Debug)]
-pub struct TuiBuilder {
-    event_tx: Sender<Event>,
-    event_rx: Receiver<Event>,
-    res_tx: UnboundedSender<DbResponse>,
-    res_rx: UnboundedReceiver<DbResponse>,
-    pub req_tx: UnboundedSender<DbRequest>,
-    req_rx: UnboundedReceiver<DbRequest>,
-}
-
 impl Tui {
     pub fn builder() -> TuiBuilder {
         TuiBuilder::default()
@@ -19,7 +9,7 @@ impl Tui {
 impl Default for TuiBuilder {
     fn default() -> Self {
         let (event_tx, event_rx) = tokio::sync::mpsc::channel(2);
-        let (res_tx, res_rx) = tokio::sync::mpsc::unbounded_channel();
+        let (res_tx, res_rx) = tokio::sync::mpsc::channel(2);
         let (req_tx, req_rx) = tokio::sync::mpsc::unbounded_channel();
 
         Self {
